requirements.txt
streamlit
joblib
numpy
streamlit_option_menu.option_menu
importlib
PIL.Image
pandas
xlsxwriter